observer: refactoring node.go to node_utils
